+++ comms_champion/app/cc_view/src/widget/field/BundleFieldWidget.cpp
@@ -92,7 +92,7 @@ void BundleFieldWidget::updatePropertiesImpl(const QVariantMap& props)
     for (auto idx = 0U; idx < count; ++idx) {
         auto* memberFieldWidget = m_members[idx];
         assert(memberFieldWidget != nullptr);
-        memberFieldWidget->updateProperties(membersProps[idx]);
+        memberFieldWidget->updateProperties(membersProps[static_cast<int>(idx)]);
     }
 }
